#7 - Integrate the form result with the backend
The signup screen currently creates a new account and then redirect to the login page

The login screen retrieves a sessionId and then navigate to the url of the patient interface with the sessionid as a param

Note: All changes is only applied in the javascript logic. The changes made in the HTML is auto-formatting.